src/requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
autopep8==1.5.4
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==2.2.15
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.1
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
pytz==2020.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
zope.interface==5.2.0
